added the exception handling code

+++ src/app/components/contacts/contact-list/contact-list.component.ts
@@ -13,6 +13,9 @@ export class ContactListComponent {
   contacts: Contact[] = [];
   selectedContact!: Contact | null;
   isFlagged: boolean = false;
+  items: any[] = [];
+  errorMessage: string | null = null;
+  
   constructor(private contactService: ContactService, private router: Router, private notificationService: NotificationService) {
     this.contactService.flag$.subscribe(value => {
       this.isFlagged = value;
@@ -21,20 +24,47 @@ export class ContactListComponent {
   
   ngOnInit(): void {
     this.contacts = this.contactService.getContacts();
+    this.getItems(); 
+  }
+
+  getItems() {
+    this.contactService.getItems().subscribe(
+      (response) => {
+        this.items = response;
+        this.errorMessage = null; // Reset error message on success
+      },
+      (error) => {
+        this.errorMessage = error; // Set error message on failure
+      }
+    );
   }
 
   deleteContact(id: number): void {
-    this.contactService.deleteContact(id);
-    this.contacts = this.contactService.getContacts();
-    this.showSuccess();
+    if(confirm('are you sure you want to delete?')){
+      this.contactService.deleteContact(id);
+      this.contacts = this.contactService.getContacts();
+      this.showSuccess();
+      this.deleteItem(123);
+    }
   }
-  
+
+  deleteItem(id: number) {
+    this.contactService.deleteItem(id).subscribe(
+      () => {
+        this.errorMessage = null; // Reset error message on success
+      },
+      (error) => {
+        this.errorMessage = error; // Set error message on failure
+      }
+    );
+  }
+
 
   editContact(contact: Contact): void {
-    this.contactService.setFlag(true);
-    
+    this.contactService.setFlag(true);    
     this.selectedContact = contact; // Set the selected contact
   }
+
   addContact(): void {
     this.router.navigate(['/create']); 
   }
